fix: gut fungi

+++ Data/Entities/User/UserCbcWAutoDiff.cs
@@ -15,9 +15,9 @@ public class UserCbcWAutoDiff
 {
     public class Consts
     {
-        public const double PlatletCountMin = 100;
-        public const double PlatletCountMax = 500;
-        public const double PlatletCountStep = 1;
+        public const int PlatletCountMin = 100;
+        public const int PlatletCountMax = 500;
+        public const int PlatletCountStep = 1;
     }
 
     [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
